requirements.txt
mysql-connector-python>=8.0.32
pandas>=1.5.3
PyYAML>=6.0